♻️ Move views into Views module
This closes #20 on GitHub.
